+++ TestTools/StreamChatTestTools/Mocks/StreamChat/Workers/MessageUpdater_Mock.swift
@@ -53,6 +53,7 @@ final class MessageUpdater_Mock: MessageUpdater {
     @Atomic var flagMessage_flag: Bool?
     @Atomic var flagMessage_messageId: MessageId?
     @Atomic var flagMessage_cid: ChannelId?
+    @Atomic var flagMessage_reason: String?
     @Atomic var flagMessage_completion: ((Error?) -> Void)?
 
     @Atomic var addReaction_type: MessageReactionType?
@@ -141,6 +142,7 @@ final class MessageUpdater_Mock: MessageUpdater {
         flagMessage_flag = nil
         flagMessage_messageId = nil
         flagMessage_cid = nil
+        flagMessage_reason = nil
         flagMessage_completion = nil
 
         addReaction_type = nil
@@ -280,10 +282,17 @@ final class MessageUpdater_Mock: MessageUpdater {
         }
     }
 
-    override func flagMessage(_ flag: Bool, with messageId: MessageId, in cid: ChannelId, completion: ((Error?) -> Void)? = nil) {
+    override func flagMessage(
+        _ flag: Bool,
+        with messageId: MessageId,
+        in cid: ChannelId,
+        reason: String? = nil,
+        completion: ((Error?) -> Void)? = nil
+    ) {
         flagMessage_flag = flag
         flagMessage_messageId = messageId
         flagMessage_cid = cid
+        flagMessage_reason = reason
         flagMessage_completion = completion
     }